refactor: deduplicate repo utility functions into shared helpers
