+++ crates/core-node-internal/src/services/repo/add.rs
@@ -1,12 +1,12 @@
 use crate::Result;
-use crate::encoding::{RepoAddRequest, RepoAddResponse, RepoSource};
+use crate::encoding::{RepoAddRequest, RepoAddResponse};
 use crate::names;
 use crate::services::repo::refresh::read_or_create_repos;
+use crate::services::repo::{json_entry_identity, repo_source_to_json};
 use config::consts::PeppyDirs;
 use peppylib::messaging::ServiceRequestContext;
 use peppylib::types::Payload;
 use peppylib::{MessengerHandle, PeppyError, PeppyResult, ServiceMessenger};
-use serde_json::Value;
 use tokio::task::JoinHandle;
 use tracing::debug;
 
@@ -49,50 +49,6 @@ async fn handle_repo_add_request(
     })
 }
 
-fn repo_source_to_json(id: u64, source: &RepoSource) -> Value {
-    let mut map = serde_json::Map::new();
-    map.insert("id".to_string(), Value::Number(id.into()));
-    match source {
-        RepoSource::Fs(path) => {
-            map.insert("type".to_string(), Value::String("fs".to_string()));
-            map.insert(
-                "path".to_string(),
-                Value::String(path.to_string_lossy().into_owned()),
-            );
-        }
-        RepoSource::Git { repo_url, repo_ref } => {
-            map.insert("type".to_string(), Value::String("git".to_string()));
-            map.insert("url".to_string(), Value::String(repo_url.clone()));
-            if let Some(r) = repo_ref {
-                map.insert("ref".to_string(), Value::String(r.to_string()));
-            }
-        }
-        RepoSource::Url(url) => {
-            map.insert("type".to_string(), Value::String("url".to_string()));
-            map.insert("url".to_string(), Value::String(url.clone()));
-        }
-    }
-    Value::Object(map)
-}
-
-/// Returns the identity string used for duplicate detection.
-fn repo_source_identity(source: &RepoSource) -> String {
-    match source {
-        RepoSource::Fs(path) => path.to_string_lossy().into_owned(),
-        RepoSource::Git { repo_url, .. } => repo_url.clone(),
-        RepoSource::Url(url) => url.clone(),
-    }
-}
-
-/// Returns the identity value from a persisted JSON entry (path for fs, url for git/url).
-fn json_entry_identity(entry: &Value) -> Option<&str> {
-    let typ = entry.get("type")?.as_str()?;
-    match typ {
-        "fs" => entry.get("path")?.as_str(),
-        _ => entry.get("url")?.as_str(),
-    }
-}
-
 fn handle_repo_add_request_inner(
     context: &ServiceRequestContext,
     peppy_dirs: &PeppyDirs,
@@ -107,7 +63,7 @@ fn handle_repo_add_request_inner(
         request.source
     );
 
-    let identity = repo_source_identity(&request.source);
+    let identity = request.source.identity();
     if identity.trim().is_empty() {
         return RepoAddResponse::failure("repository path/URL must not be empty").encode();
     }
@@ -121,7 +77,6 @@ fn handle_repo_add_request_inner(
         Err(e) => return RepoAddResponse::failure(e.to_string()).encode(),
     };
 
-    // Duplicate check
     let new_identity = identity.trim();
     let is_duplicate = repos
         .iter()
@@ -132,7 +87,6 @@ fn handle_repo_add_request_inner(
             .encode();
     }
 
-    // Compute the next available id
     let next_id = repos
         .iter()
         .filter_map(|e| e.get("id").and_then(|v| v.as_u64()))
@@ -140,7 +94,6 @@ fn handle_repo_add_request_inner(
         .map(|max| max + 1)
         .unwrap_or(1);
 
-    // Append and write back (JSON is valid JSON5, use pretty for user readability)
     repos.push(repo_source_to_json(next_id, &request.source));
     let content = serde_json::to_string_pretty(&repos)
         .map_err(|e| crate::Error::Encoding(format!("failed to serialize repositories: {e}")))?;
